refactor: Setup 통합 — onboarding 5단계 → 4단계 (#17)
QA 피드백: init 프로세스 너무 길음. calendar / policies / notifications
세 화면이 모두 "AI 추론 confirm" 패턴이라 한 화면으로 묶음.

새 흐름
- intro → goal-intake (1/4 목표) → goal-classify (2/4 분류) → setup (3/4
  마무리) → weekly-plan (4/4 계획) → morning-brief → today
- 이전 7단계 → 5단계 (intro·morning-brief 포함)

SetupScreen
- 한 화면 안에 3 섹션: 일정(캘린더 + 고정 일정) · 지킬 시간(수면/점심/break)
  · 알림(모닝·저녁 칩 + 사전 알림 토글)
- 답 콜백: "인터뷰 답에서 수면 23:00–07:00 로 잡고…" 헤더 부근에 노출
- CTA 하나 "다 좋아요" — fixedSchedules.create + notifications.updateSettings
  를 한 번에 호출하지 않고 변경분만 patch (저장 실패해도 흐름은 진행)

정리
- SetupProgress 를 src/components/SetupProgress.tsx 로 추출 (이전엔
  CalendarScheduleScreen 안에서 export). 모든 단계 화면이 같은 컴포넌트.
- ScreenId 에서 'calendar-schedule', 'policies-notifications' 제거,
  'setup' 추가. STATE_TO_SCREEN 의 ONBOARDING_CALENDAR/MANUAL_SCHEDULE/
  POLICIES/NOTIFICATIONS 모두 'setup' 으로 매핑.
- NAV_META 갱신 (back 체인 정리).
- 모든 단계 progress 를 X/4 로 재카운트.
- CalendarScheduleScreen.tsx / PoliciesNotificationsScreen.tsx 삭제.
