+++ src/app/ReActionMerged.tsx
@@ -4,9 +4,8 @@ import { MergedTabBar } from '../components/TabBar';
 import { SystemIntroScreen } from '../screens/SystemIntroScreen';
 import { GoalIntakeScreen } from '../screens/GoalIntakeScreen';
 import { GoalClassificationScreen } from '../screens/GoalClassificationScreen';
-import { CalendarScheduleScreen } from '../screens/CalendarScheduleScreen';
+import { SetupScreen } from '../screens/SetupScreen';
 import { WeeklyPlanGenerationScreen } from '../screens/WeeklyPlanGenerationScreen';
-import { PoliciesNotificationsScreen } from '../screens/PoliciesNotificationsScreen';
 import { MorningBriefScreen } from '../screens/MorningBriefScreen';
 import { InboxScreen } from '../screens/InboxScreen';
 import { SettingsScreen } from '../screens/SettingsScreen';
@@ -29,10 +28,9 @@ const NAV_META: Record<ScreenId, { label: string; back: ScreenId | null }> = {
   'intro':                  { label: 'RE:ACTION',      back: null },
   'goal-intake':            { label: '목표 파악',      back: 'intro' },
   'goal-classify':          { label: '목표 분류',      back: 'goal-intake' },
-  'calendar-schedule':      { label: '고정 일정',      back: 'goal-classify' },
-  'weekly-plan':            { label: '주간 계획 생성', back: 'calendar-schedule' },
-  'policies-notifications': { label: '마무리 확인',    back: 'weekly-plan' },
-  'morning-brief':          { label: '모닝 브리프',    back: 'policies-notifications' },
+  'setup':                  { label: '마무리 확인',    back: 'goal-classify' },
+  'weekly-plan':            { label: '주간 계획 생성', back: 'setup' },
+  'morning-brief':          { label: '모닝 브리프',    back: 'weekly-plan' },
   'today':                  { label: '오늘의 실행',    back: null },
   'focus':                  { label: '집중 모드',      back: 'today' },
   'recovery':               { label: '복구 코치',      back: 'today' },
@@ -167,16 +165,13 @@ export function ReActionMerged({ hideTabs = false }: ReActionMergedProps) {
           <GoalIntakeScreen onDone={() => setScreen('goal-classify')} />
         )}
         {screen === 'goal-classify' && (
-          <GoalClassificationScreen onNext={() => setScreen('calendar-schedule')} />
+          <GoalClassificationScreen onNext={() => setScreen('setup')} />
         )}
-        {screen === 'calendar-schedule' && (
-          <CalendarScheduleScreen onNext={() => setScreen('weekly-plan')} />
+        {screen === 'setup' && (
+          <SetupScreen onDone={() => setScreen('weekly-plan')} />
         )}
         {screen === 'weekly-plan' && (
-          <WeeklyPlanGenerationScreen onContinue={() => setScreen('policies-notifications')} />
-        )}
-        {screen === 'policies-notifications' && (
-          <PoliciesNotificationsScreen onDone={() => setScreen('morning-brief')} />
+          <WeeklyPlanGenerationScreen onContinue={() => setScreen('morning-brief')} />
         )}
         {screen === 'morning-brief' && (
           <MorningBriefScreen onStart={() => { setTab('today'); setScreen('today'); }} />
