+++ src/types/index.ts
@@ -18,9 +18,8 @@ export type ScreenId =
   | 'intro'
   | 'goal-intake'
   | 'goal-classify'
-  | 'calendar-schedule'
+  | 'setup'
   | 'weekly-plan'
-  | 'policies-notifications'
   | 'morning-brief'
   | 'today'
   | 'focus'
